+++ src/Core/CoreComponents.cpp
@@ -201,7 +201,7 @@ RenderLayer::~RenderLayer()
 
 bool RenderLayer::m_dirtyOrder = false;
 
-MoveCollider2Points::MoveCollider2Points(const Vector2<float> &offset_) :
+MoveCollider2Points::MoveCollider2Points(const Vector2<int> &offset_) :
     m_offset(offset_)
 {
 }
