+++ src/Core/CoreComponents.h
@@ -228,10 +228,10 @@ private:
 
 struct MoveCollider2Points
 {
-    MoveCollider2Points(const Vector2<float> &offset_);
-    Vector2<float> m_point1;
-    Vector2<float> m_point2;
-    const Vector2<float> m_offset;
+    MoveCollider2Points(const Vector2<int> &offset_);
+    Vector2<int> m_point1;
+    Vector2<int> m_point2;
+    const Vector2<int> m_offset;
     FrameTimer<false> m_timer;
 };
